fixed ceph health monitor aligned with PR#1571 fixed an error with string.lstrip; fixed way to use TimeoutSampler fixed code review comments: moved some functions from test file to ocp.py` Signed-off-by: aviadp <apolak@redhat.com>
fixed an error with string.lstrip; fixed way to use TimeoutSampler


fixed code review comments: moved some functions from test file to ocp.py
